backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
numpy
openai
requests 
python-multipart